Issue #109 support 'then' in compose is ready
