+++ kaspresso/kaspresso-framework/src/main/kotlin/com/kaspersky/kaspresso/compose/pack/ActionsOnElementsPack.kt
@@ -32,9 +32,8 @@ class ActionsOnElementsPack {
     fun <Type> or(element: Type, action: Type.() -> Unit): ComplexComposeBranchBuilder<Type>
             where Type : BaseActions, Type : BaseAssertions,
                   Type : Interceptable<ViewInteraction, ViewAssertion, ViewAction> {
-        return ComplexComposeBranchBuilder(element, { action.invoke(element) }).also {
-            complexComposeBranchBuilders += it
-        }
+        return ComplexComposeBranchBuilder(element, { action.invoke(element) })
+            .also { complexComposeBranchBuilders += it }
     }
 
     /**
@@ -47,9 +46,8 @@ class ActionsOnElementsPack {
     fun <Type> or(element: Type, action: Type.() -> Unit): ComplexComposeBranchBuilder<Type>
             where Type : UiBaseActions, Type : UiBaseAssertions,
                   Type : UiInterceptable<UiObjectInteraction, UiObjectAssertion, UiObjectAction> {
-        return ComplexComposeBranchBuilder(element, { action.invoke(element) }).also {
-            complexComposeBranchBuilders += it
-        }
+        return ComplexComposeBranchBuilder(element, { action.invoke(element) })
+            .also { complexComposeBranchBuilders += it }
     }
 
     /**
